requirements.txt
openai
replicate
requests
Pillow
